+++ src/block-fly/display/levelControls.ts
@@ -1,5 +1,43 @@
 import * as Promise from "bluebird";
 
+export function bindLevelsControls(callback: (levels: string[]) => void): void {
+  let loadRemoteButton = document.getElementById("load-remote");
+  loadRemoteButton.addEventListener("click", () => {
+      let url = (document.getElementById("remote-path") as HTMLInputElement).value;
+
+      if (url !== "") {
+        getXHRLevels(url).then(callback)
+        .catch((error: string): void => {
+          alert(error);
+        });
+      } else {
+        alert("Please enter an URL");
+      }
+  });
+
+  let loadDefaultsButton = document.getElementById("load-defaults");
+  loadDefaultsButton.addEventListener("click", () => {
+    getDefaultLevels().then(callback)
+    .catch((error: string): void => {
+      alert(error);
+    });
+  });
+
+  let loadLocalLevels = document.getElementById("load-file");
+  loadLocalLevels.addEventListener("click", () => {
+      let levelsFile = (document.getElementById("local-levels") as HTMLInputElement).files[0];
+
+      if (levelsFile !== undefined) {
+        getLevelsFromFile(levelsFile).then(callback)
+        .catch((error: string): void => {
+          alert(error);
+        });
+      } else {
+        alert("Please choose a file");
+      }
+  });
+}
+
 export function getDefaultLevels(): Promise<string[]> {
   return getXHRLevels("assets/default-levels.json");
 }
